separate redux store and slice
