+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
-import { store } from './redux'
+import { store } from './redux/store'
 import { Provider } from 'react-redux'
 
 import { ChakraProvider } from '@chakra-ui/react'
