+++ src/redux/store.ts
@@ -1,0 +1,12 @@
+import { configureStore } from '@reduxjs/toolkit'
+import { userExposureReducer } from './userExposureSlice'
+
+export const store = configureStore({
+  reducer: {
+    userExposure: userExposureReducer,
+  },
+})
+
+export type RootState = ReturnType<typeof store.getState>
+
+export type AppDispatch = typeof store.dispatch
